backend-java/cv-yolo/python/requirements.txt
torch>=1.7
torchvision
Pillow
